refactor(result): dto package로 이동
